yoyiyi

+++ app/src/main/java/com/yoyiyi/soleil/bean/app/video/VideoDetailComment.kt
@@ -144,7 +144,7 @@ data class VideoDetailComment(var code: Int,
                 data class PendantBean(var pid: Int,
                                        var name: String,
                                        var image: String,
-                                       var expire: Int) {
+                                       var expire: Long) {
                     /**
                      * pid : 0
                      * name :
@@ -314,7 +314,7 @@ data class VideoDetailComment(var code: Int,
                 data class PendantBeanX(var pid: Int,
                                         var name: String,
                                         var image: String,
-                                        var expire: Int) {
+                                        var expire: Long) {
                     /**
                      * pid : 0
                      * name :
